renaming plugin, setting package name, fixing package exports, allowing multi-site configuration in package
