+++ packages/headless/.eslintrc.js
