+++ packages/headless/.gitignore
@@ -1,0 +1,5 @@
+node_modules
+dist
+*.js
+*.d.ts
+!.eslintrc.js
