+++ packages/headless/.npmignore
@@ -1,0 +1,5 @@
+src/
+.eslintrc.js
+.gitignore
+node_modules
+tsconfig.json
