+++ packages/headless/src/index.ts
@@ -1,0 +1,5 @@
+export { gql, useQuery } from '@apollo/client';
+export { ApiProvider } from './Provider';
+export { useClient } from './client';
+export * from './middleware';
+export * from './hooks';
