--- packages/wp-api/src/index.ts
@@ -1,5 +1,0 @@
-export { gql, useQuery } from '@apollo/client';
-export { ApiProvider } from './Provider';
-export { config } from './client';
-export * as middelware from './middleware';
-export * as hooks from './hooks';
