+++ plugins/wpe-headless/class.wpe-headless.php
@@ -3,7 +3,7 @@
  * Class for handling Headless Wordpress
  *
  */
-class WP_Headless {
+class WPE_Headless {
 	/**
 	 * Add various hooks.
 	 *
@@ -15,39 +15,39 @@ class WP_Headless {
 
         add_filter('preview_post_link', array( __CLASS__, 'set_post_preview_link' ));
         add_filter('post_link', array( __CLASS__, 'set_post_link' ));
-        WP_Headless_Api::init();
-        WP_Headless_Redirect::init();
+        WPE_Headless_Api::init();
+        WPE_Headless_Redirect::init();
     }
 
     public static function activate() {
         flush_rewrite_rules();
 
-        $secret_key = WP_Headless_Constants::get_secret_key_option();
+        $secret_key = WPE_Headless_Constants::get_secret_key_option();
 
         if(isset($secret_key)) {
-            update_option(WP_Headless_Constants::SECRET_KEY, wp_generate_uuid4());
+            update_option(WPE_Headless_Constants::SECRET_KEY, wp_generate_uuid4());
         } else {
-            add_option(WP_Headless_Constants::SECRET_KEY, wp_generate_uuid4());
+            add_option(WPE_Headless_Constants::SECRET_KEY, wp_generate_uuid4());
         }
 
-        $frontend_uri = WP_Headless_Constants::get_frontend_uri_option();
+        $frontend_uri = WPE_Headless_Constants::get_frontend_uri_option();
 
         if(isset($frontend_uri)) {
-            update_option(WP_Headless_Constants::FRONTEND_URI, '');
+            update_option(WPE_Headless_Constants::FRONTEND_URI, '');
         } else {
-            add_option(WP_Headless_Constants::FRONTEND_URI, '');
+            add_option(WPE_Headless_Constants::FRONTEND_URI, '');
         }
     }
 
     public static function set_post_preview_link() {
-        $base_uri = WP_Headless_Constants::get_frontend_uri_option();
+        $base_uri = WPE_Headless_Constants::get_frontend_uri_option();
         $post = get_post();
 
         return $base_uri . base64_encode('post:' . $post->ID) . '/?status=' . $post->post_status . '&preview=true';
     }
 
     public static function set_post_link() {
-        $base_uri = WP_Headless_Constants::get_frontend_uri_option();
+        $base_uri = WPE_Headless_Constants::get_frontend_uri_option();
         $post = get_post();
 
         if ($post->post_status === 'draft') {
@@ -60,37 +60,37 @@ class WP_Headless {
     public static function deactivate() {
         flush_rewrite_rules();
 
-        delete_option(WP_Headless_Constants::SECRET_KEY);
-        delete_option(WP_Headless_Constants::FRONTEND_URI);
+        delete_option(WPE_Headless_Constants::SECRET_KEY);
+        delete_option(WPE_Headless_Constants::FRONTEND_URI);
 
         remove_filter( 'preview_post_link', array( __CLASS__, 'set_post_preview_link' ) );
     }
 
     public static function admin_init() {
         // register a new setting for "general" page
-        register_setting('general', WP_Headless_Constants::SECRET_KEY);
-        register_setting('general', WP_Headless_Constants::FRONTEND_URI);
+        register_setting('general', WPE_Headless_Constants::SECRET_KEY);
+        register_setting('general', WPE_Headless_Constants::FRONTEND_URI);
 
         // register a new section in the "general" page
         add_settings_section(
-            'wp_Headlesss_settings',
+            'WPE_Headlesss_settings',
             'WP Authentication Codes', array( __CLASS__, 'settings_section_callback' ),
             'general'
         );
 
-        // register a new field in the "wp_Headlesss_settings" section, inside the "general" page
+        // register a new field in the "WPE_Headlesss_settings" section, inside the "general" page
         add_settings_field(
-            WP_Headless_Constants::SECRET_KEY,
+            WPE_Headless_Constants::SECRET_KEY,
             'API Secret', array( __CLASS__, 'settings_api_secret_callback' ),
             'general',
-            'wp_Headlesss_settings'
+            'WPE_Headlesss_settings'
         );
 
         add_settings_field(
-            WP_Headless_Constants::FRONTEND_URI,
+            WPE_Headless_Constants::FRONTEND_URI,
             'Preview Base Address (URL)', array( __CLASS__, 'settings_base_uri_callback' ),
             'general',
-            'wp_Headlesss_settings'
+            'WPE_Headlesss_settings'
         );
     }
 
@@ -102,24 +102,24 @@ class WP_Headless {
     // section content cb
     public static function settings_api_secret_callback() {
         // get the value of the setting we've registered with register_setting()
-        $setting = WP_Headless_Constants::get_secret_key_option();
+        $setting = WPE_Headless_Constants::get_secret_key_option();
         // output the field
         ?>
 
 
         <input type="text" disabled class="regular-text" value="<?php echo isset( $setting ) ? esc_attr( $setting ) : ''; ?>">
-        <input type="hidden" class="regular-text" name="<?php echo WP_Headless_Constants::SECRET_KEY ?>" value="<?php echo isset( $setting ) ? esc_attr( $setting ) : ''; ?>">
+        <input type="hidden" class="regular-text" name="<?php echo WPE_Headless_Constants::SECRET_KEY ?>" value="<?php echo isset( $setting ) ? esc_attr( $setting ) : ''; ?>">
         <?php
     }
 
     // section content cb
     public static function settings_base_uri_callback() {
         // get the value of the setting we've registered with register_setting()
-        $setting = WP_Headless_Constants::get_frontend_uri_option();
+        $setting = WPE_Headless_Constants::get_frontend_uri_option();
         // output the field
         ?>
 
-        <input type="text" class="regular-text" name="<?php echo WP_Headless_Constants::FRONTEND_URI ?>" value="<?php echo $setting ?>">
+        <input type="text" class="regular-text" name="<?php echo WPE_Headless_Constants::FRONTEND_URI ?>" value="<?php echo $setting ?>">
         <?php
     }
 }
